+++ tamago-gateway/src/main/kotlin/tamago/server/gateway/presentation/letter/v1/controller/LetterController.kt
@@ -1,17 +1,14 @@
 package tamago.server.gateway.presentation.letter.v1.controller
 
 import io.swagger.v3.oas.annotations.Parameter
-import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.*
-import tamago.server.core.letter.LetterCommandUseCase
 import tamago.server.core.letter.LetterFacade
 import tamago.server.core.letter.LetterQueryUseCase
 import tamago.server.core.letter.domain.vo.ReceivedLetterId
 import tamago.server.core.user.domain.aggregate.User
 import tamago.server.gateway.presentation.letter.v1.api.LetterApi
-import tamago.server.gateway.presentation.letter.v1.request.LetterCreateRequest
 import tamago.server.gateway.presentation.letter.v1.response.LetterResponse
 import tamago.server.gateway.common.response.CustomResponse
 import tamago.server.gateway.common.annotation.CurrentUser
@@ -19,7 +16,6 @@ import tamago.server.gateway.common.annotation.CurrentUser
 @RestController
 class LetterController(
     private val letterQueryUseCase: LetterQueryUseCase,
-    private val letterCommandUseCase: LetterCommandUseCase,
     private val letterFacade: LetterFacade,
 ) : LetterApi {
 
@@ -49,15 +45,4 @@ class LetterController(
         return CustomResponse.noContent()
     }
 
-    @ResponseStatus(HttpStatus.CREATED)
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @PostMapping("/api/v1/letters")
-    override fun createTemplate(@RequestBody @Valid request: LetterCreateRequest): CustomResponse<Void> {
-        letterCommandUseCase.createTemplate(
-            title = request.title,
-            content = request.content,
-        )
-
-        return CustomResponse.created()
-    }
 }
